refactor(libs): moved queries helpers to libs
make it accessible for all services
